Use tuples for color specification (#9149)
In the Matplotlib 'Specifying Colors' section, neither list nor array is a valid color specification type.  When this example is run using Python 3.5, and matplotlib 2.0.2 (latest), it errors with the message: 

"ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()"

This change preserves the semantics of the example while allowing it to successfully run by converting to tuples for the matplotlib color specification.